Multiple languages
I decided to add multiple languages and then drop down to the navigation bar/header so that it'd be more accessible to everybody. It should be SEO friendly. I did this because I saw that somebody from France used it, so I'm like, "Oh well, and Sherod uses it too. So I might as well add that French language, English just because it's a worldwide language that everybody uses."