Language Update and Image Optimization.
I switched the formats of the about page images from JPG to WEBP. I also changed the dimensions a little bit and added Jarod's image. Also did a bunch of fixing for the untranslated parts of the recipes and contact page. Also updated the hero section with the logo that Roman made.